Splitting app.js, not done
