+++ Server/Pages/callerobj.mjs
@@ -1,0 +1,15 @@
+import * as fs from 'fs'
+
+import { getPostData } from "../helpers.mjs";
+import { addCaller } from '../helpers.mjs';
+
+export function page_callerobj(req, res, path) {
+    // Creates a date object 'd' the fs.writeFileSync uses to name it's documents by date
+    // and writes the stringified json to its respective json document in the Server/ServerData/CallerDB/caller-year-month-day.
+    getPostData(req).then(caller => {
+        // If the file does not exist, it will instead create one that is ready for json object input
+        if (!fs.existsSync(path)) exportObject(path, '[]', res);
+        //Check if an error has occured, if it hasn't: End. To prevent writing to ended stream
+        if (!addCaller(path, caller, res)) res.end('ok');
+    })
+}
