+++ Server/ServerData/CallerDB/callers-2022-3-21.json
@@ -1,0 +1,78 @@
+[
+    {
+        "name": "Thomas",
+        "ReportedLocation": {
+            "address": "Brandevej 10, Aalborg",
+            "lat": 57.046833,
+            "lng": 9.913824
+        },
+        "situation": "Fire",
+        "number": 11471317,
+        "timeset": "21.4.2022 10.08.15",
+        "AMLLocation": {
+            "lat": 57.046832,
+            "lng": 9.913825
+        },
+        "injuries": "at least 4",
+        "answered": false,
+        "answering": true,
+        "useful": true,
+        "description": "I'm dead dammit",
+        "id": "3e8e9055-4261-4f86-bb28-fb46593eab8b"
+    },
+    {
+        "name": "Thomæs",
+        "address": "Unknown address",
+        "situation": "unknown situation",
+        "number": 22256521,
+        "timeset": "21.4.2022 10.47.44",
+        "AMLLocation": {
+            "lat": 57.052578,
+            "lng": 9.911738
+        },
+        "injuries": "unknown injuries",
+        "answered": false,
+        "answering": true,
+        "useful": true,
+        "description": "No description provided",
+        "id": "0726b2b7-7147-49a3-867f-69607ad1a0ab"
+    },
+    {
+        "name": "ASd",
+        "address": "Sigrid Undsets Vej 210B",
+        "situation": "fdg",
+        "number": 40017116,
+        "timeset": "21.4.2022 10.48.57",
+        "AMLLocation": {
+            "lat": 57.052578,
+            "lng": 9.911738
+        },
+        "injuries": "unknown injuries",
+        "answered": false,
+        "answering": true,
+        "useful": true,
+        "description": "No description provided",
+        "id": "0e2dca0f-aca3-4125-9760-48446f99cebf"
+    },
+    {
+        "name": "ad",
+        "location": {
+            "address": "Unknown address",
+            "lat": "",
+            "lng": ""
+        },
+        "situation": "unknown situation",
+        "number": 11157595,
+        "timeset": "21.4.2022 16.16.32",
+        "AMLLocation": {
+            "lat": 57.046832,
+            "lng": 9.913825
+        },
+        "injuries": "unknown injuries",
+        "answered": false,
+        "answering": false,
+        "useful": true,
+        "description": "No description provided",
+        "id": "690d0144-58f2-43af-a2ca-74fbc3ddfb1b"
+    }
+]
