+++ package.json
@@ -2,5 +2,6 @@
   "dependencies": {
     "socket.io": "^4.4.1",
     "uuid": "^8.3.2"
-  }
+  },
+  "type": "module"
 }
